simplify and fix dockerfiles
The dockerfiles had some complexity issues which made them harder to
read and debug. This patch attempts to simplify them.

The main reason for this patch however, is that command-executer was
being built outside the containers, on the local system, and moving it
to the container didn't work correctly, as the glibc version could
be different.

I have also tried adding the multithreaded test to the tests to be run,
but it doesn't seem to correctly run under command-executer and I didn't
get too deep into what the issue it might have is. I think it's not
critical for testing purposes and will come back to it later. It seems
to work fine without it being wrapped by command-executer.

Signed-off-by: Petre Eftime <epetre@amazon.com>